% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacom_sim_fx.R
\name{C_f}
\alias{C_f}
\title{Metacommunity process: colonization}
\usage{
C_f(
  v = "Sum interactions",
  d_c = "Sensitivity to interactions",
  c_0 = "Colonization at zero interactions",
  c_max = "Colonization at max interactions"
)
}
\arguments{
\item{v}{sum of interactions}

\item{d_c}{sensitivity to interactions}

\item{c_0}{colonization at zero interactions}

\item{c_max}{colonization at max interactions}
}
\description{
The total effect of ecological interactions on the colonization probabilty
}
