% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\docType{data}
\name{orig-no-seed-E}
\alias{orig-no-seed-E}
\title{Environmental variable for 1000 sites}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 1 columns.
}
\usage{
orig-no-seed-E
}
\description{
This vector contains the environmental variable used for the simulations and analysis in the 2021 paper, 'The Internal Structure of Metacommunities'
}
\keyword{datasets}
