% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orig-no-seed-MEMsel.R
\docType{data}
\name{orig-no-seed-MEMsel}
\alias{orig-no-seed-MEMsel}
\title{Selected MEMs for 1000 sites}
\format{
An object of class \code{orthobasisSp} (inherits from \code{orthobasis}, \code{data.frame}) with 1000 rows and 78 columns.
}
\usage{
orig-no-seed-MEMsel
}
\description{
This dataset has the 78 variables used in the spatial component of the analysis, significant variables from Moran Eigenvector Maps used in analysis in the 2021 paper, 'The Internal Structure of Metacommunities'
}
\keyword{datasets}
