% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pars_fx.R
\name{prep_pars}
\alias{prep_pars}
\title{Prepare metacommunity parameters}
\usage{
prep_pars(
  N = NULL,
  D = NULL,
  R = NULL,
  nicheOpt = NULL,
  breadth = NULL,
  alpha = NULL,
  interx_col = NULL,
  interx_ext = NULL,
  makeRDS = FALSE,
  whereToSave = NULL,
  objName = NULL
)
}
\arguments{
\item{N}{number of patches for the simulation}

\item{D}{number of environmental variables simulated}

\item{R}{number of species for the simulation}

\item{nicheOpt}{vector with niche optima for each species. Default is range equally divided among species.}

\item{breadth}{niche breadth. As reference 0.8 as narrow and 2 as broad with a quadratic response.}

\item{alpha}{dispersal parameter associated to an exponential distribution}

\item{makeRDS}{whether to save the output as an RDS file}

\item{whereToSave}{file path to save output}

\item{objName}{name for the output}

\item{interex_col}{Species sensitivity to interactions. For our simulations, sensitivity is set the same for all species, and both for colonization and extinction}

\item{interex_ext}{Species sensitivity to interactions. For our simulations, sensitivity is set the same for all species, and both for colonization and extinction}
}
\description{
This function is used to prepare the parameters in a useful format and avoid lengthy code for every simulation. It will output all parameters in a list ready for the main_sim_fx()
}
