% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_processing_fx.R
\name{sites_plot}
\alias{sites_plot}
\title{Base Plots: Sites}
\usage{
sites_plot(data, plotMain = NULL, colorVar = NULL, colorLegend = "none")
}
\arguments{
\item{data}{formatted data from get_sites_data()}

\item{plotMain}{main title for the figure}

\item{colorVar}{variable for the color variation. Color by species or by niche optima}

\item{colorLegend}{should the color legend be included, what is the title?}
}
\description{
Base species plot
}
