% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metaco.R
\name{sum_interactions}
\alias{sum_interactions}
\title{Metacommunity process: ecological interactions}
\usage{
sum_interactions(
  A = "Species interactions matrix",
  Y = "Species occurrence matrix"
)
}
\arguments{
\item{A}{species interaction matrix}

\item{Y}{species occurrence matrix}
}
\description{
Sum of ecological interactions. Computes the sum of ecological interactions for every location and every species
}
\keyword{ecological}
\keyword{interactions}
