% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XY.R
\docType{data}
\name{XY}
\alias{XY}
\title{Coordinates variable for 1000 sites}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 2 columns.
}
\usage{
XY
}
\description{
This data has the xy coordinates used for the simulations and analysis in the 2021 paper, 'The Internal Structure of Metacommunities'
}
\keyword{datasets}
